HW08 исправления по замечаниям
