+++ HW08/src/main/java/ru/otus/hw/repositories/CommentRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface CommentRepository extends MongoRepository<Comment,String> {
     List<Comment> findByBookId(String bookId);
 
-    void deleteByBookId (String bookId);
+    void deleteAllByBookId (String bookId);
 }
