+++ HW08/src/main/java/ru/otus/hw/repositories/events/CommentDeleteByBookDeleteEvent.java
@@ -1,0 +1,22 @@
+package ru.otus.hw.repositories.events;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.mongodb.core.mapping.event.AbstractMongoEventListener;
+import org.springframework.data.mongodb.core.mapping.event.AfterDeleteEvent;
+import org.springframework.stereotype.Component;
+import ru.otus.hw.models.Book;
+import ru.otus.hw.repositories.CommentRepository;
+
+@Component
+@RequiredArgsConstructor
+public class CommentDeleteByBookDeleteEvent extends AbstractMongoEventListener<Book> {
+    private final CommentRepository commentRepository;
+
+    @Override
+    public void onAfterDelete(AfterDeleteEvent<Book> event) {
+        super.onAfterDelete(event);
+        var deletedBook = event.getSource();
+        var idBook = deletedBook.get("_id").toString();
+        commentRepository.deleteAllByBookId(idBook);
+    }
+}
