+++ HW08/src/main/java/ru/otus/hw/services/BookServiceImpl.java
@@ -9,7 +9,6 @@ import ru.otus.hw.models.Book;
 import ru.otus.hw.models.Genre;
 import ru.otus.hw.repositories.AuthorRepository;
 import ru.otus.hw.repositories.BookRepository;
-import ru.otus.hw.repositories.CommentRepository;
 import ru.otus.hw.repositories.GenreRepository;
 
 import java.util.List;
@@ -27,9 +26,6 @@ public class BookServiceImpl implements BookService {
 
     private final BookRepository bookRepository;
 
-    private final CommentRepository commentRepository;
-
-
     @Transactional(readOnly = true)
     @Override
     public Optional<Book> findById(String id) {
@@ -45,7 +41,7 @@ public class BookServiceImpl implements BookService {
     @Transactional
     @Override
     public Book insert(String title, String authorId, List<String> genreIds) {
-        var book = new Book(title, getAuthorById(authorId), getGenresByIds(genreIds));;
+        var book = new Book(title, getAuthorById(authorId), getGenresByIds(genreIds));
         return bookRepository.save(book);
     }
 
@@ -61,10 +57,8 @@ public class BookServiceImpl implements BookService {
     @Override
     @Transactional
     public void deleteById(String id) {
-        commentRepository.deleteByBookId(id);
         bookRepository.deleteById(id);
     }
-
 
     private Author getAuthorById(String authorId) {
         return authorRepository.findById(authorId)
